==> 01_book_data_collection.ipynb <==
# Book Ratings Data Collection & Cleaning

### Import libraries and datasets

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from scipy import sparse

#import raw data
books = pd.read_csv('./data/raw/user_rating_0_to_1000.csv')
books.head(2)

#set 1-5 values for user reviews
books['Rating'].value_counts()

books['Rating'].replace('it was amazing', 5, inplace=True)
books['Rating'].replace('really liked it', 4, inplace=True)
books['Rating'].replace('liked it', 3, inplace=True)
books['Rating'].replace('it was ok', 2, inplace=True)
books['Rating'].replace('did not like it', 1, inplace=True)
books['Rating'].replace("This user doesn't have any rating", 0, inplace=True)

books['Rating'].value_counts()

#create book ratings dataframe for app
# https://stackoverflow.com/questions/42350029/assign-a-number-to-each-unique-value-in-a-list
book_names = books['Name']
dictionary = {i: indi for indi, i in enumerate(set(book_names))}
numbers = [dictionary[n] for n in book_names]
books['book_id'] = numbers
books.sort_values(by=['book_id'])

#create recommender dataframe
recommender_df = books[['book_id', 'Name', 'ID', 'Rating']]
recommender_df.rename(columns={'ID': 'user_id', 'Name': 'book_name', 'Rating': 'rating'}, inplace=True)
#reorder for streamlit app
recommender_df = recomender_df[['book_name', 'book_id', 'user_id', 'rating']]

recommender_df.head(2)

#export dataframe
recommender_df.to_csv('./data/clean/books_df.csv')
