==> 01c_music_recommender.ipynb <==
# Music Dataset Recommender

### Import Libraries and data

import pandas as pd
import numpy as np
import spacy
import en_core_web_lg
import string
import pickle
from textblob import TextBlob
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

# #note: pitchfork.csv cleaned in Google Colab due to NLTK import issues to create pitchfork_cleaned
#using the following functions:

# # modified from: https://stackoverflow.com/questions/5511708/adding-words-to-nltk-stoplist
# stop_words = nltk.corpus.stopwords.words('english')
# additional_stopwords = ["“", "’", "”", "."]
# stop_words.extend(additional_stopwords)

# #function to take each review and set it as lowercase and remove stop words/punctuation
# def sentence_cleaner(sentence):
#     cleaner = set(stop_words + list(string.punctuation))
#     new_sentence = [i for i in word_tokenize(sentence.lower()) if i not in cleaner]
#     new_sentence = " ".join(new_sentence)
#     return new_sentence

# cleaned_list = []
# for i in range(len(music_df['content'])):
#     processed_sentence = sentence_cleaner(music_df['content'][i])
#     cleaned_list.append(processed_sentence)
    
# #create cleaned review dataframe
# cleaned_content_df = pd.DataFrame(cleaned_list, columns=['cleaned_content'])

# #dataframe with original text, cleaned text
# music_df_cleaned = music_df.join(cleaned_content_df, how='outer')

# #export dataframe
# music_df_cleaned.to_csv('pitchfork_cleaned.csv')

music_df = pd.read_csv('./data/clean/pitchfork_cleaned.csv', index_col=0)

music_df.shape

music_df.head(1)

music_df.drop_duplicates(subset=['content'], inplace=True)
music_df.dropna(subset=['artist'], inplace=True)
music_df.dropna(subset=['title'], inplace=True)
music_df.dropna(subset=['cleaned_content'], inplace=True)
music_df.reset_index(inplace=True)

### Data Cleaning

nlp = spacy.load('en_core_web_lg')

music_df['cleaned_content'].isnull().sum()

Note: these cells previous run and output as "spacy_list"

# spacy_list = []
# for i in range(len(music_df['content'])):
#     spacy_object = nlp(music_df['cleaned_content'][i])
#     spacy_list.append(spacy_object)
#     print(f'{i}')

# #save spacy_list as txt file to load in streamlit
# # https://stackoverflow.com/questions/899103/writing-a-list-to-a-file-with-python/899176
# with open('./streamlit_app/data/spacy_list.txt', 'w', encoding="utf-8") as f:
#     for item in spacy_list:
#         f.write("%s\n" % item)

# open txt file as spacy_list
# https://stackoverflow.com/questions/30523943/how-to-read-just-the-first-column-of-each-row-of-a-csv-file
with open('./data/clean/spacy_list.txt', encoding="utf-8") as f:
    spacy_list = [line.split(',')[0] for line in f]

# music_df.to_pickle('./streamlit_app/data/music.pk1')

### Create Recommender Function

#recomender with genre filter
def recomender(input):
    recomender_list = []
    genre = music_df['genre'][input]
    for i in range(len(spacy_list)):
        vector = nlp(spacy_list[i]).similarity(nlp(spacy_list[input]))
        recomender_list.append(vector)
        recomender_df = pd.DataFrame(recomender_list)
        recomender_df['artist'] = music_df['artist']
        recomender_df['title'] = music_df['title']
        recomender_df['genre'] = music_df['genre']
        recomender_df = recomender_df.sort_values(by=0, ascending=False)
        recomender_df = recomender_df[recomender_df['genre'] == genre]
    return recomender_df[0:6]

recomender(2171)

recomender(4278)

recomender(3391)

recomender(18286)

recomender(6523)

### Add polarity and subjectivity to recommender

analyzer = SentimentIntensityAnalyzer()
sentiment_list = []
polarity_list = []
for i in range(len(music_df)):
    text = music_df['cleaned_content'][i]
    sentiment_score = TextBlob(text).sentiment[0]
    polarity_score = analyzer.polarity_scores(text)['compound']
    sentiment_list.append(sentiment_score)
    polarity_list.append(polarity_score)

sentiment_df = pd.DataFrame(sentiment_list, columns=['sentiment_score'])
polarity_df = pd.DataFrame(polarity_list, columns=['polarity_score'])
music_df = music_df.join(sentiment_df, how='outer')
music_df = music_df.join(polarity_df, how='outer')

#recomender with genre filter + polarity and sentiment engineered features
def recomender_sentiment(input):
    recomender_list = []
    genre = music_df['genre'][input]
    for i in range(len(spacy_list)):
        vector = nlp(spacy_list[i]).similarity(nlp(spacy_list[input]))
        recomender_list.append(vector)
        recomender_df = pd.DataFrame(recomender_list)
        recomender_df['artist'] = music_df['artist']
        recomender_df['title'] = music_df['title']
        recomender_df['genre'] = music_df['genre']
        recomender_df['sentiment_score'] = music_df['sentiment_score']
        recomender_df['polarity_score'] = music_df['polarity_score']
        recomender_df['engineered_rec'] = recomender_df[0] + recomender_df['sentiment_score'] + recomender_df['polarity_score']
        recomender_df = recomender_df.sort_values(by='engineered_rec', ascending=False)
        recomender_df = recomender_df[recomender_df['genre'] == genre]
    return recomender_df[0:6]

# #return index value for artist name
# def get_artist_name(artist)
#     which(sapply(music_df, function(x) any(x == "artist")))

recomender_sentiment(2171)

recomender_sentiment(4278)
