==> 02c_book_recommender.ipynb <==
# Book Recommender

### Import libraries and Data

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from scipy import sparse
from sklearn.metrics.pairwise import pairwise_distances, cosine_distances, cosine_similarity

books_df = pd.read_csv('./data/clean/books_df.csv', index_col=0)

books_df.head(1)

### Prepare recommender using cosine distances

#pivot table from the books_df dataframe
pivot_df = pd.pivot_table(books_df, index='book_name', columns='user_id', values='rating')
#create sparse matrix
sparse_df = sparse.csr_matrix(pivot_df.fillna(0))
#calculate cosine distances for similarities
recommender = pairwise_distances(sparse_df, metric='cosine')
#export as a dataframe
rec_df = pd.DataFrame(recommender, columns=pivot_df.index, index=pivot_df.index)

### Define book Recommender function

def book_recommender(book_title):
    return 1- rec_df[book_title].sort_values()[1:6]

### Example Recommendations

#Using the recommender function, input a book title to give cosine similarities, sorted
book_recommender('To Kill a Mockingbird')

Similar to "To Kill a Mockingbird" on Goodreads.com:
- The Great Gatsby
- The Catcher in the Rye
- Animal Farm
- Lord of the Flies

book_recommender('1984')

Similar to "1984" on Goodreads.com:
- Catcher in the Rye
- Fahrenheit 451
- Brave New World
- Lord of the Flies

book_recommender("The Fellowship of the Ring (The Lord of the Rings, #1)")

Similar to "The Fellowship of the Ring" on Goodreads.com:
- The Two Towers
- The Hobbit, or There and Back Again

book_recommender('Life of Pi')

Similar to "Life of Pi" on Goodreads.com:
- The Kite Runner
- Memoirs of a Geisha
- The Curious Incident of the Dog in the Night-Time
- Water for Elephants

### Export for streamlit app

#add book name, re-arrange for streamlit app
rec_df.insert(0, "book_name", rec_df.index)

# save rec_df for use in streamlit app
rec_df.to_pickle('./streamlit_app/data/books.pk1')
