==> 03a_game_data_collection.ipynb <==
# Board Game Ratings Data Collection & Cleaning

### Import libraries and datasets

import pandas as pd
import numpy as np
from sklearn.preprocessing import LabelEncoder

Data available at:
    https://www.kaggle.com/jvanelteren/boardgamegeek-reviews

### Import data with game information

#import raw data - 1st csv
games = pd.read_csv('./data/raw/games_detailed_info.csv', index_col=0)
games.rename(columns={'id':'ID'}, inplace=True)
games.head(2)

columns_to_drop = ["boardgameintegration", "boardgamecompilation", "Party Game Rank", 
"Abstract Game Rank", "Thematic Rank", "War Game Rank", "Customizable Rank", 
"Children's Game Rank", "RPG Item Rank", "Accessory Rank", "Video Game Rank", 
"Amiga Rank", "Commodore 64 Rank", "Arcade Rank", "Atari ST Rank", "Family Game Rank", 
"Strategy Game Rank", "boardgameartist", "boardgameimplementation", "boardgameexpansion", 
"boardgamefamily", "alternate", "stddev", "owned", "trading", "wanting", "wishing", "numcomments",
"numweights", "averageweight", "suggested_playerage", "suggested_language_dependence",
"boardgamecategory", "bayesaverage", "Board Game Rank", "median", "boardgamedesigner",
"yearpublished", "thumbnail", "image", "boardgamemechanic", "type", "primary", "suggested_num_players",
"boardgamepublisher", "minplaytime", "maxplaytime", 'minage', 'description']
games.drop(columns=columns_to_drop, inplace=True)

### Import data with user reviews

#import raw data - 2nd csv
games_2 = pd.read_csv('./data/raw/bgg-15m-reviews.csv', index_col=0)

#anonymize data using Label Encoder
#modified from: https://stackoverflow.com/questions/48008334/anonymize-specific-columns-with-pii-in-pandas-dataframe-python
label_encoder = LabelEncoder()
games_2['user'] = games_2['user'].astype(str)
user_anonymzed = label_encoder.fit(games_2['user']).transform(games_2['user'])
games_2['user'] = user_anonymzed

games_2.head(2)

### Create dataframe with game information and user reviews

#create merged dataframe
games_df = pd.merge(games_2, games, on='ID')

games_df.head(3)

games_df.shape

### Create subset (5M) of larger 15M dataframe and export dataframe

#take smaller sample size to use in streamlit app
games_df_subset = games_df.sample(n=5_000_000, replace=False)

#export to csv
games_df_subset.to_csv('./data/clean/games_df.csv')
