==> 03c_game_recommender.ipynb <==
# Board Game Recommender

### Import libraries and Data

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from scipy import sparse
from sklearn.metrics.pairwise import pairwise_distances, cosine_distances, cosine_similarity

### Import Full dataframe (5M Reviews)

games_df_full = pd.read_csv('./data/clean/games_df.csv', index_col=0)

games_df_full.head(1)

games_df_full.shape

### Create a subset of data for recomender due to memory issues/file sizes

# subset of full dataset for recommender
games_df = games_df_full[['name', 'ID', 'user', 'rating']]

#Due to memory issues when creating pivot tables below, reduce the size of the dataframe
#Keep only games with more than 500 reviews
#modified from: https://stackoverflow.com/questions/51079543/pandas-groupby-apply-vs-transform-with-specific-functions
games_df_subset = games_df[games_df.groupby(['name'])['rating'].transform('count') >= 500]
games_df_subset.shape

#Sample 100,000 reviews
games_df_subset = games_df_subset.sample(n=100000, replace=False, random_state=123)
games_df_subset.shape

games_df_subset.head(2)

games_df_subset.user.nunique()

games_df_subset.name.nunique()

### Prepare recommender using cosine distances

#pivot table from the books_df dataframe
pivot_df = pd.pivot_table(games_df_subset, index='name', columns='user', values='rating')
#create sparse matrix
sparse_df = sparse.csr_matrix(pivot_df.fillna(0))
#calculate cosine distances for similarities
recommender = pairwise_distances(sparse_df, metric='cosine')
#export as a dataframe
rec_df = pd.DataFrame(recommender, columns=pivot_df.index, index=pivot_df.index)

# get average rating, number of players, playing time, etc to filter by in app, group by game ID
games_info = games_df_full.groupby('name').mean()
games_info.drop(columns=['user'], inplace=True)
games_info.rename(columns={'rating': 'average_rating'}, inplace=True)
games_info.head(2)

rec_df = pd.merge(left = rec_df, right= games_info, right_index = True, left_index = True)

### Define book Recommender function

def game_recommender(game_title, average_rating = 7.5, minplayers = -np.inf, playingtime = np.inf):
    '''Users can input a game title, specify what the average rating should be above, specify minplayers and playingtime'''
    rec_df2 = rec_df[rec_df['average_rating'] >= average_rating]
    rec_df3 = rec_df2[rec_df2['minplayers'] >= minplayers]
    rec_df4 = rec_df3[rec_df3['playingtime'] <= playingtime]
    return 1- rec_df4[game_title].sort_values()[1:6]

### Example Recommendations

game_recommender('Monopoly')

- No similar titles in "Fans also like" section of boardgamegeek.com

game_recommender('Terraforming Mars')

Similar titles in "Fans also like" section of boardgamegeek.com:
- Wingspan

game_recommender('7 Wonders')

Similar titles in "Fans also like" section of boardgamegeek.com:
- Dominion

game_recommender('Cosmic Encounter')

- No similar titles in "Fans also like" section of boardgamegeek.com

Accounting for limited similarities in boardgamegeek's webiste for "fans also like":
- Dataset had to be scaled down considerably due to memory issues and file size.
- Recommender does not take into account type of board game - future work may include adding a board game "style" filter.

### Export for streamlit app

#add game title, re-arrange for streamlit app
rec_df.insert(0, "game_title", rec_df.index)

# save rec_df for use in streamlit app
rec_df.to_pickle('./streamlit_app/data/games.pk1')
