==> 04b_music_user_rev_recommender.ipynb <==
# Music Album Recommender

### Import libraries and Data

import pandas as pd
import numpy as np
from scipy import sparse
from sklearn.metrics.pairwise import pairwise_distances, cosine_distances, cosine_similarity

music_df_full = pd.read_csv('./data/clean/amazon_music_reviews.csv')

music_df_full.drop(columns=['Unnamed: 0', 'level_0', 'index'], inplace=True)

music_df_full.shape

music_df_full.head(1)

music_df_subset = music_df_full[music_df_full.groupby(['product_title'])['review_body'].transform('count') >= 50]

# subset of full dataset for recommender
music_df = music_df_subset[['product_title', 'review_id', 'customer_id', 'star_rating', 'review_body']]
music_df = music_df.sample(n=100_000)

### Prepare recommender using cosine distances

#pivot table from the books_df dataframe
pivot_df = pd.pivot_table(music_df, index='product_title', columns='customer_id', values='star_rating')
#create sparse matrix
sparse_df = sparse.csr_matrix(pivot_df.fillna(0))
#calculate cosine distances for similarities
recommender = pairwise_distances(sparse_df, metric='cosine')
#export as a dataframe
rec_df = pd.DataFrame(recommender, columns=pivot_df.index, index=pivot_df.index)

# get average rating, number of players, playing time, etc to filter by in app, group by game ID
music_info = music_df_full.groupby('product_title').mean()
music_info.rename(columns={'star_rating': 'average_rating'}, inplace=True)
music_info.drop(columns=['customer_id'])

rec_df = pd.merge(left = rec_df, right= music_info, right_index = True, left_index = True)
rec_df.insert(0, 'album_name', rec_df.index)

### Define book Recommender function

def album_recommender(album_title):
    return 1- rec_df[album_title].sort_values()[1:6]

album_recommender("Room for Squares")

album_recommender("Random Access Memories")

# save rec_df for use in streamlit app
# rec_df.to_pickle('./streamlit_app/data/albums.pk1')
